move DirectiveInfo to graphql.schema.idl package
